Move NUIPSenTestKitAssertions to NUIParseTests
The fact that this header relies on the SenTesingKit framwork, leads me to the
conclusion, that it is not meant to be used anywhere else than in unit testing
targets.
It defines UnitTest assertions, that to my knowledge are only used in
NUIPSTAssertionsTests.m

Because this file was caught by the source_files pattern of the podspec file,
it is included in the umbrella headers generated by cocoapods when applying the
new 'use_framworks!' option.

This leads to an error, because you normally wouldn't link the framework itself
against the SenTesingKit framework, but it is automatically imported.

I thus moved it to the unit test target folder, where IMHO it belongs anyways.